+++ src/main/java/roomescape/Controller/PageController.java
@@ -1,13 +1,18 @@
-package roomescape;
+package roomescape.Controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class HomeController {
+public class PageController {
 
     @GetMapping("/")
     public String home() {
         return "home";
+    }
+
+    @GetMapping("/reservation")
+    public String reservationPage() {
+        return "reservation";
     }
 }
