+++ src/main/java/roomescape/Domain/Reservation.java
@@ -1,0 +1,42 @@
+package roomescape.Domain;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+public class Reservation {
+
+    private final Integer id;
+    private final String name;
+    private final LocalDate date;
+    private final LocalTime time;
+
+    public Reservation(Integer id, String name, LocalDate date, LocalTime time) {
+        this.id = id;
+        this.name = name;
+        this.date = date;
+        this.time = time;
+    }
+
+    public Reservation() {
+        this.id = null;
+        this.name = null;
+        this.date = null;
+        this.time = null;
+    }
+
+    public Integer getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public LocalDate getDate() {
+        return date;
+    }
+
+    public LocalTime getTime() {
+        return time;
+    }
+}
